# backend/src/main/resources/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "saving",
      "description": "All about your savings"
    }
  ],
  "paths": {
    "/api/v1/savings/{savingId}": {
      "get": {
        "tags": [
          "saving"
        ],
        "summary": "Get saving by ID",
        "description": "Returns a single saving",
        "operationId": "getSavingById",
        "parameters": [
          {
            "name": "savingId",
            "in": "path",
            "description": "ID of saving to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saving found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Saving not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "saving"
        ],
        "summary": "Update an existing saving by ID",
        "operationId": "updateSaving",
        "parameters": [
          {
            "name": "savingId",
            "in": "path",
            "description": "ID of saving to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated saving",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Saving not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Saving updated",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "ID mismatch",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "saving"
        ],
        "summary": "Delete a saving by ID",
        "operationId": "deleteSaving",
        "parameters": [
          {
            "name": "savingId",
            "in": "path",
            "description": "ID of saving to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Saving not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "200": {
            "description": "Saving updated",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/savings": {
      "get": {
        "tags": [
          "saving"
        ],
        "summary": "Get all savings",
        "description": "Returns all available savings",
        "operationId": "getAllSavings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "saving"
        ],
        "summary": "Add a new saving",
        "operationId": "addSaving",
        "requestBody": {
          "description": "Saving to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingDTO"
              },
              "examples": {
                "SavingExample": {
                  "description": "SavingExample",
                  "value": {
                    "id": 10,
                    "amount": 10000.0,
                    "date": "2025-09-04T10:47:00+00:00",
                    "goal": "Apartament",
                    "description": "My dream apartament"
                  }
                },
                "BadExample": {
                  "description": "BadExample",
                  "value": {
                    "id": 1,
                    "amount": -2000.0,
                    "date": "2222-09-04T10:47:00+00:00",
                    "goal": "",
                    "description": "bad description"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Saving added successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/test": {
      "get": {
        "tags": [
          "test-controller"
        ],
        "operationId": "testEndpoint",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SavingDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "goal": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
}